LayoutTests/fast/xpath/xpath-detached-import-assert-expected.txt

A test case from https://bugs.webkit.org/show_bug.cgi?id=84793. Test passes if it doesn't hit ASSERT(parents.size() >= depth + 1).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result.numberValue is NaN
PASS successfullyParsed is true

TEST COMPLETE
